refactor: 그룹 멤버 조회/멤버십 검증 책임을 GroupService로 이동
